WIP: beginning to rewrite msaa ppe
