+++ ccb/frm/framebuffer.cpp
